Fix package name to match the repository name
If the git repo is cloned to Sublime Text Packages folder (and thus the new directory
will be named after the repo), the paths to the settings files must match the directory.

This change means that any settings users have customised (in "Settings - User")
will be on the wrong path (<packages>/CFBeautifier) and need to be recreated or
moved to the new path (<packages>/CFBeautifier).